+++ server/src/main/java/com/jaywright/SpringTest/SpringTestApplication.java
