+++ server/src/main/java/com/jaywright/SpringTest/controller/MainController.java
@@ -1,0 +1,13 @@
+package com.jaywright.SpringTest.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MainController {
+
+    @GetMapping("/test")
+    public String getTest() {
+        return "This is a test";
+    }
+}
